# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/product-categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 100
    put:
      operationId: updateAdminProductCategory
      tags: [admin]
      summary: Save or clear a product category image
      security:
        - firebaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCategoryImageInput"
      responses:
        "200":
          description: Category image saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductCategory"
        "400":
          description: Invalid category image
        "401":
          description: Authentication required
        "404":
          description: Product category not found
...
components:
...
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: FirebaseIDToken
  schemas:
...
    ProductCategoryImageInput:
      type: object
      properties:
        categoryImagePath:
          type: [string, "null"]
          maxLength: 500
        categoryImageAlt:
          type: [string, "null"]
          maxLength: 180
      required: [categoryImagePath, categoryImageAlt]
    ProductCategory:
      type: object
      properties:
        categoryId:
          type: string
        categoryName:
          type: string
        categoryImagePath:
          type: [string, "null"]
        categoryImageUrl:
          type: [string, "null"]
        categoryImageAlt:
          type: [string, "null"]
        active:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [categoryId, categoryName, categoryImagePath, categoryImageUrl, categoryImageAlt, active, displayOrder, createdAt, updatedAt]
...